// src/courier-tracking/constant/group.d.ts
declare const Group: {
    OrderPlaced: 'OrderPlaced';
    Dispatched: 'Dispatched';
    InTransit: 'InTransit';
    OutForDelivery: 'OutForDelivery';
    FailedDelivery: 'FailedDelivery';
    Delivered: 'Delivered';
    Lost: 'Lost';
    Damaged: 'Damaged';
    Returned: 'Returned';
};
export { Group };