// src/courier-tracking/constant/stage.d.ts
declare const Stage: {
    AWBGenerated: 'AWBGenerated';
    OrderPlaced: 'OrderPlaced';
    PickupPending: 'PickupPending';
    PickupFailed: 'PickupFailed';
    PickedUp: 'PickedUp';
    OutForPickup: 'OutForPickup';
    InTransit: 'InTransit';
    OutForDelivery: 'OutForDelivery';
    NotServiceable: 'NotServiceable';
    Delivered: 'Delivered';
    DeliveryFailed: 'DeliveryFailed';
    OrderCancelled: 'OrderCancelled';
    ShipmentDelayed: 'ShipmentDelayed';
    RTORequested: 'RTORequested';
    RTO: 'RTO';
    RTOInTransit: 'RTOInTransit';
    RTOOutForDelivery: 'RTOOutForDelivery';
    RTODelivered: 'RTODelivered';
    RTODeliveryFailed: 'RTODeliveryFailed';
    RTOShipmentDelayed: 'RTOShipmentDelayed';
    ReturnOrderPlaced: 'ReturnOrderPlaced';
    ExchangePickUp: 'ExchangePickUp';
    ExchangeInTransit: 'ExchangeInTransit';
    ExchangeDelivered: 'ExchangeDelivered';
    Lost: 'Lost';
    Damaged: 'Damaged';
    ShipmentHeld: 'ShipmentHeld';
};
export { Stage };